Refactor: Extract Nanoc::Int::Props
